sat: initial revision
Initial revision of a minimal sat solver. This is a naive solver without
any DPLL optimizations; the approach is basically to try every possible
option.

Clauses are implemented as reference counted persistent unbalanced
dynamic trees. Literals are just integer values, with a negative value
indicating the complement of the variable.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>