sat: do not unwind on satisfiability
Once unit propagation and pure literal elimination are introduced,
variable assignments must be unwound on failure; however, since variable
assignments must be printed at the end, they should not be unwound on
success.

Change unwind() so it is only called on failure, and a final call to
unwind(0) will unwind all clauses.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>