sat: use an array for clause backtracking
Clause backtracking has an upper bound of the number of literals it
contains + 1; allocate an array to hold all of these, and use pointer
arithmetic to "allocate" and "deallocate" blocks from this array instead
of malloc() and free(). This also eliminates a linked list from the
unwind() fastpath, giving ~10% performance boost.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>